pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src", "utils"]

[project]
name = "GENETIS-ARA"
version = "0.1"
description="GENETIS Genetic Algorithm to evolve antennas in ARA deployment"
dependencies = [
    #"shared-code @ git+https://github.com/osu-particle-astrophysics/Shared-Code.git@add_hpol", 
    "numpy",
    "pandas",
    "matplotlib",
    "black", # using for formatting + documentation
    "isort", # also formatting
    "flake8", # linter
    "pyyaml"
]